♻️ #17 recordを整備した
